* master: (66 commits)
  provider/aws: Fix dependency violation when deleting Internet Gateways
  command/remote-config: failing tests
  update CHANGELOG
  command/remote-config: do a pull with `terraform remote config`
  command/remote-{pull,push}: colorize and show success output
  command/remote-config: lowercase the type so that Atlas works, for example
  command/remote-config: show flag parse errors
  command/remote-config: remove weird error case that shows no error message
  command: when setting up state, only write back if local is newer
  minor code cleanups to get acceptance tests passing
  update CHANGELOG
  providers/digitalocean: add dot in GET response
  providers/digitalocean: force fqdn in dns rr value
  update CHANGELOG
  small code cleanup
  Add proper reading/updating of tags for S3
  provider/aws: Add tags to S3
  Documentation for ASG Tags added
  Tags support added for AWS ASG
  command/output: don't panic if no root module in state [GH-1263]
  ...

+++ builtin/providers/cloudstack/resource_cloudstack_vpn_customer_gateway.go
@@ -87,11 +87,11 @@ func resourceCloudStackVPNCustomerGatewayCreate(d *schema.ResourceData, meta int
 	}
 
 	if esplifetime, ok := d.GetOk("esp_lifetime"); ok {
-		p.SetEsplifetime(esplifetime.(int))
+		p.SetEsplifetime(int64(esplifetime.(int)))
 	}
 
 	if ikelifetime, ok := d.GetOk("ike_lifetime"); ok {
-		p.SetIkelifetime(ikelifetime.(int))
+		p.SetIkelifetime(int64(ikelifetime.(int)))
 	}
 
 	// Create the new VPN Customer Gateway
@@ -128,8 +128,8 @@ func resourceCloudStackVPNCustomerGatewayRead(d *schema.ResourceData, meta inter
 	d.Set("ike_policy", v.Ikepolicy)
 	d.Set("ipsec_psk", v.Ipsecpsk)
 	d.Set("dpd", v.Dpd)
-	d.Set("esp_lifetime", v.Esplifetime)
-	d.Set("ike_lifetime", v.Ikelifetime)
+	d.Set("esp_lifetime", int(v.Esplifetime))
+	d.Set("ike_lifetime", int(v.Ikelifetime))
 
 	return nil
 }
@@ -154,11 +154,11 @@ func resourceCloudStackVPNCustomerGatewayUpdate(d *schema.ResourceData, meta int
 	}
 
 	if esplifetime, ok := d.GetOk("esp_lifetime"); ok {
-		p.SetEsplifetime(esplifetime.(int))
+		p.SetEsplifetime(int64(esplifetime.(int)))
 	}
 
 	if ikelifetime, ok := d.GetOk("ike_lifetime"); ok {
-		p.SetIkelifetime(ikelifetime.(int))
+		p.SetIkelifetime(int64(ikelifetime.(int)))
 	}
 
 	// Update the VPN Customer Gateway
